NA映画8.0抽奖器
版本号V1.7

+++ 代码/handle.h
@@ -3,8 +3,9 @@
 #include "dialog.h"
 #include <QMessageBox>
 #include <random>
+#include <QDebug>
 
-typedef long long ll;
+typedef long ll;
 
 DWORD WINAPI adadadad(LPVOID Wparam);
